+++ packages/link/Link.stories.tsx
@@ -1,5 +1,6 @@
-import { StoryFn, Meta } from '@storybook/react'
-import { Link, LinkProps } from '.'
+import { Story, Meta } from '@storybook/react'
+import Link from './Link'
+import { LinkProps } from './types'
 
 export default {
   component: Link,
@@ -14,6 +15,6 @@ export default {
       description: 'Fade visited links',
     },
   },
-} satisfies Meta
+} as Meta
 
-export const Basic: StoryFn<LinkProps> = (props) => <Link {...props} />
+export const Basic: Story<LinkProps> = (props) => <Link {...props} />
