--- packages/link/LinkStyles.module.css
@@ -1,14 +1,0 @@
-.link {
-  text-decoration: none;
-  color: var(--lido-color-primary);
-
-  :hover {
-    color: var(--lido-color-primaryHover);
-  }
-}
-
-.link.fadedVisited {
-  &:visited {
-    color: var(--lido-color-primaryVisited);
-  }
-}
