+++ packages/link/index.ts
@@ -1,1 +1,2 @@
-export * from './Link'
+export { default as Link } from './Link'
+export * from './types'
