+++ packages/link/types.ts
@@ -1,0 +1,4 @@
+import { LidoComponentProps } from '../utils'
+export type { Theme } from '../theme'
+
+export type LinkProps = LidoComponentProps<'a'> & { fadeVisited?: boolean }
